[ADD] Added a category saga and reducer
