+++ frontend/src/reducers/category.js
@@ -1,0 +1,19 @@
+const defaultState = {
+    addCategory: false,
+  };
+  
+  const category = (state = defaultState, action) => {
+    switch (action.type) {
+      case 'ADD_CATEGORY_DEMAND': return {
+        addCategory: !state.addCategory
+      }
+      default:
+        return {
+          ...state,
+        };
+    }
+  };
+  
+  export default category;
+  
+
