+++ frontend/src/reducers/index.js
@@ -5,9 +5,10 @@ import message from './message';
 import project from './project';
 import demand from './demand';
 import task from "./task";
+import category from './category'
 
 const reducers = combineReducers({
-  user, message, project, demand, task
+  user, message, project, demand, task, category
 });
 
 export default reducers;
