+++ frontend/src/sagas/projectSaga.js
@@ -28,9 +28,6 @@ function* handleProjectUpdateDemand(action){
     yield put({type: "PROJECTUPDATEFORM_DEMAND", projectID: action.projectID, projectName:action.projectName });
 }
 
-function *handleCategoryCreationDemand(){
-    yield put({type: 'ADD_CATEGORY_DEMAND'});
-}
 
  function *projectSaga(){
     yield takeEvery('USER_IS_PROJECT_DEMAND', handleProjectDemand);
@@ -38,7 +35,6 @@ function *handleCategoryCreationDemand(){
     yield takeEvery('USER_PROJECTFORM_DEMAND', handleAddProjectDemand);
     yield takeEvery('USER_VIEW_PROJECTTASKS', handleViewProjectTasks);
     yield takeEvery('USER_PROJECTUPDATE_DEMAND', handleProjectUpdateDemand);
-    yield takeEvery('USER_ADD_CATEGORY_DEMAND', handleCategoryCreationDemand);
 }
 
 export default projectSaga;
